# api-specs/swagger/personCollectionApi.yaml
openapi: 3.0.3
info:
  title: SOA Lab 1
  description: Вариант 8602. Алексей Волков & Сандов Кирилл.
  version: 1.0.0
...
servers:
  - url: https://localhost:28444/central-service/api
    description: central-service
  - url: https://localhost:8444/api/proxy
    description: proxy-service
paths:
...
  /demography/hair-color/{hairColor}:
    servers:
      - url: https://localhost:8444/api/proxy
        description: proxy-service
    get:
      tags:
        - proxy-service
      summary: Количество людей по цвету волос
      parameters:
        - name: hairColor
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Количество людей
          content:
            application/json:
              schema:
                type: integer
        "400": { description: Неверный цвет волос }
  /demography/nationality/{nationality}/eye-color/{eyeColor}/percentage:
    servers:
      - url: https://localhost:8444/api/proxy
        description: proxy-service
    get:
      tags:
        - proxy-service
      summary: Доля людей по цвету глаз и национальности
      parameters:
        - name: nationality
          in: path
          required: true
          schema:
            type: string
            enum:
              - RUSSIA
              - CHINA
              - INDIA
              - ITALY
              - SOUTH_KOREA
        - name: eyeColor
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Процент людей
          content:
            application/json:
              schema:
                type: number
                format: double
        "400": { description: Неверные значения параметров }
...